Conflicts:
	build/jam/packages/Haiku
	headers/os/kernel/OS.h
	headers/os/opengl/GLRenderer.h
	headers/private/shared/cpu_type.h
	src/add-ons/kernel/drivers/power/acpi_battery/acpi_battery.h
	src/bin/sysinfo.cpp
	src/bin/top.c
	src/system/kernel/arch/x86/arch_system_info.cpp
	src/system/kernel/port.cpp

+++ src/apps/diskprobe/TypeEditors.cpp
@@ -187,7 +187,7 @@ TypeEditorView::TypeMatches()
 	system_info info;
 	get_system_info(&info);
 
-	return fEditor.FileSize() / B_PAGE_SIZE < info.max_pages / 2;
+	return uint64(fEditor.FileSize()) / B_PAGE_SIZE < info.max_pages / 2;
 }
